chore: move permissions definition to actions folder
also update examples to handle permissions in array
