+++ examples/next/pages/ui/components/storage/storage-browser/default-auth/routed/[locations]/[location-detail]/index.page.tsx
@@ -17,6 +17,10 @@ export default function Page() {
 
   const { path, ...location } = query;
 
+  if (typeof location.permissions === 'string') {
+    location.permissions = [location.permissions];
+  }
+
   return (
     <Flex>
       <Button
