+++ examples/next/pages/ui/components/storage/storage-browser/managed-auth/routed/[locations]/[location-detail]/index.page.tsx
@@ -15,6 +15,10 @@ export default function Page() {
 
   const { path, ...location } = query;
 
+  if (typeof location.permissions === 'string') {
+    location.permissions = [location.permissions];
+  }
+
   return (
     <Flex>
       <SignOutButton
