+++ packages/react-storage/src/components/StorageBrowser/credentials/store/store.ts
@@ -1,4 +1,4 @@
-import { LocationPermissions } from '../../credentials/types';
+import { LocationPermissions } from '../../actions';
 import {
   AWSTemporaryCredentials,
   StorageValidationErrorCode,
